+++ src/components/List/components/public/ListFilterPreset/ListFilterPreset.component.tsx
@@ -7,7 +7,7 @@ import { useListFilterPreset } from "./ListFilterPreset.utils";
 import { Badge, type BadgeProps } from "@/components/Badge";
 import { Text } from "@/components/Text";
 import { useTestId } from "@/utils/props.utils";
-import { normalizeToPaletteColor } from "@utils";
+import { normalizeToPaletteColor, useUikitTranslation } from "@utils";
 
 import type {
   ListFilterPresetProps,
@@ -19,12 +19,14 @@ import type { PaletteColor } from "@types";
  * A component that displays a single filter preset button
  * and allows interacting with it.
  *
- * @version 0.0.2
+ * @version 0.0.3
  *
  * @param {ListFilterPresetProps} props
  * @returns {Nullable<JSX.Element>}
  */
-export const ListFilterPreset = (props: ListFilterPresetProps) => {
+export const ListFilterPreset = <TItem extends object = object>(
+  props: ListFilterPresetProps<TItem>
+) => {
   const { filterPreset, styleProps } = useListFilterPreset(props);
   const testId = useTestId("filter-preset-button", {
     testId: filterPreset.testId,
@@ -39,10 +41,19 @@ export const ListFilterPreset = (props: ListFilterPresetProps) => {
   );
 
   const textColor = useMemo<PaletteColor>(() => {
+    if (filterPreset.disabled) return "gray-600";
     if (filterPreset.active)
       return normalizeToPaletteColor(filterPreset.paletteColor, "dark");
     return "gray-800";
-  }, [filterPreset.active, filterPreset.paletteColor]);
+  }, [filterPreset.active, filterPreset.disabled, filterPreset.paletteColor]);
+
+  const tl = useUikitTranslation();
+
+  const title = useMemo(() => {
+    return tl.action[filterPreset.active ? "unselect" : "select"](
+      filterPreset.label
+    );
+  }, [filterPreset.active, filterPreset.label, tl]);
 
   if (filterPreset.hidden) return null;
 
@@ -52,6 +63,7 @@ export const ListFilterPreset = (props: ListFilterPresetProps) => {
       onClick={filterPreset.toggle}
       {...styleProps}
       data-testid={testId}
+      title={title}
     >
       <Text
         weight="medium"